requested changes: navigation code moved to Agent.py
